requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
docutils==0.16
idna==2.9
invoke==1.4.1
requests==2.23.0
robotframework==3.1.2
robotframework-pageobjectlibrary==1.0.2
robotframework-seleniumlibrary==4.3.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
tqdm==4.44.1
urllib3==1.25.8
